Added adaptor design pattern implementation in java alogn with refactor other design patterns
