+++ src/main/java/Other/PilotPen.java
@@ -1,0 +1,16 @@
+package Other;
+
+public class PilotPen {
+
+    public void mark(String str){
+
+
+        //
+        //
+        //
+        //
+
+
+        System.out.println(str);
+    }
+}
