+++ src/main/java/adapterDesignPattern/AssignmentWork.java
@@ -1,0 +1,18 @@
+package adapterDesignPattern;
+
+public class AssignmentWork {
+
+    private Pen p;
+
+    public Pen getP(){
+        return p;
+    }
+
+    public void setPen(Pen p){
+        this.p = p;
+    }
+
+    public void writesAssignment(String str){
+
+    }
+}
