+++ src/main/java/adapterDesignPattern/Pen.java
@@ -1,0 +1,6 @@
+package adapterDesignPattern;
+
+public interface Pen {
+
+    public void write(String str);
+}
