+++ src/main/java/adapterDesignPattern/PenAdaptor.java
@@ -1,0 +1,13 @@
+package adapterDesignPattern;
+
+import Other.PilotPen;
+
+public class PenAdaptor implements  Pen{
+
+    PilotPen pp = new PilotPen();
+
+    @Override
+    public void write(String str) {
+        pp.mark(str);
+    }
+}
