+++ src/main/java/factoryDesignPattern/OS.java
@@ -1,4 +1,4 @@
-package phone;
+package factoryDesignPattern;
 
 public interface OS {
     void spec();
