+++ src/main/java/factoryDesignPattern/OperatingSystemFactory.java
@@ -1,4 +1,4 @@
-package phone;
+package factoryDesignPattern;
 
 public class OperatingSystemFactory {
