+++ src/main/java/factoryDesignPattern/Windows.java
@@ -1,4 +1,4 @@
-package phone;
+package factoryDesignPattern;
 
 public class Windows  implements  OS{
     @Override
