--- src/main/java/phone/Android.java
@@ -1,9 +1,0 @@
-package phone;
-
-public class Android implements  OS{
-
-    @Override
-    public void spec() {
-        System.out.println("Most powerful OS!");
-    }
-}
